More compartmentalization of bundle-based-runner only utilities.
